refactor: dark mode added
